+++ ngx-numeric-range-form-field/projects/ngx-numeric-range-form-field/src/lib/form/model/numeric-range-field.model.ts
@@ -1,4 +1,12 @@
-export interface INumericRange {
-  minimum: number;
-  maximum: number;
-}
+import { FormControl, FormGroup } from '@angular/forms';
+
+export type INumericRange = {
+	minimum: number;
+	maximum: number;
+};
+
+type ControlsOf<T extends Record<string, unknown>> = {
+	[K in keyof T]: FormControl<T[K]>;
+};
+
+export type NumericRangeFormGroup = FormGroup<ControlsOf<INumericRange>>;